feat: rename Entity\MigratedTrait
