+++ src/Entity/Gateway/Charge.php
@@ -5,7 +5,7 @@ namespace App\Entity\Gateway;
 use App\Entity\Accounting\Accounting;
 use App\Entity\Accounting\Transaction;
 use App\Entity\EmbeddableMoney as Money;
-use App\Entity\Trait\MigratedEntity;
+use App\Entity\MigratedTrait;
 use App\Entity\Trait\TimestampedCreationEntity;
 use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Gateway\ChargeStatus;
@@ -30,7 +30,7 @@ use Symfony\Component\Validator\Constraints as Assert;
 #[ORM\Entity(repositoryClass: ChargeRepository::class)]
 class Charge
 {
-    use MigratedEntity;
+    use MigratedTrait;
     use TimestampedCreationEntity;
     use TimestampedUpdationEntity;
