+++ src/Entity/Project/BudgetItem.php
@@ -5,7 +5,7 @@ namespace App\Entity\Project;
 use App\Entity\EmbeddableMoney as Money;
 use App\Entity\LocalizedInterface;
 use App\Entity\LocalizedTrait;
-use App\Entity\Trait\MigratedEntity;
+use App\Entity\MigratedTrait;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\BudgetItemRepository;
 use AutoMapper\Attribute\MapProvider;
@@ -18,7 +18,7 @@ use Gedmo\Mapping\Annotation as Gedmo;
 #[ORM\Entity(repositoryClass: BudgetItemRepository::class)]
 class BudgetItem implements LocalizedInterface
 {
-    use MigratedEntity;
+    use MigratedTrait;
     use LocalizedTrait;
 
     #[ORM\Id]
