+++ src/Entity/Project/Reward.php
@@ -5,7 +5,7 @@ namespace App\Entity\Project;
 use App\Entity\EmbeddableMoney as Money;
 use App\Entity\LocalizedInterface;
 use App\Entity\LocalizedTrait;
-use App\Entity\Trait\MigratedEntity;
+use App\Entity\MigratedTrait;
 use App\Entity\Trait\TimestampedCreationEntity;
 use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Mapping\Provider\EntityMapProvider;
@@ -25,7 +25,7 @@ use Gedmo\Mapping\Annotation as Gedmo;
 #[ORM\Entity(repositoryClass: RewardRepository::class)]
 class Reward implements LocalizedInterface
 {
-    use MigratedEntity;
+    use MigratedTrait;
     use LocalizedTrait;
     use TimestampedCreationEntity;
     use TimestampedUpdationEntity;
