+++ src/Entity/User/User.php
@@ -5,8 +5,8 @@ namespace App\Entity\User;
 use App\Entity\Accounting\Accounting;
 use App\Entity\DedupedTrait;
 use App\Entity\Interface\AccountingOwnerInterface;
+use App\Entity\MigratedTrait;
 use App\Entity\Project\Project;
-use App\Entity\Trait\MigratedEntity;
 use App\Entity\Trait\TimestampedCreationEntity;
 use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Mapping\Provider\EntityMapProvider;
@@ -35,7 +35,7 @@ use Symfony\Component\Security\Core\User\UserInterface;
 #[UniqueEntity('handle', message: 'This handle is already in use.')]
 class User implements UserInterface, PasswordAuthenticatedUserInterface, AccountingOwnerInterface
 {
-    use MigratedEntity;
+    use MigratedTrait;
     use DedupedTrait;
     use TimestampedCreationEntity;
     use TimestampedUpdationEntity;
